+++ src/main/java/com/ll/quizzle/domain/quiz/dto/generation/QuizDTO.java
@@ -1,0 +1,16 @@
+package com.ll.quizzle.domain.quiz.dto.generation;
+
+import com.ll.quizzle.domain.quiz.enums.AnswerType;
+import com.ll.quizzle.domain.quiz.enums.Difficulty;
+import com.ll.quizzle.domain.quiz.enums.MainCategory;
+import com.ll.quizzle.domain.quiz.enums.SubCategory;
+
+public record QuizDTO(
+        Long id,
+        MainCategory mainCategory,
+        SubCategory subCategory,
+        AnswerType answerType,
+        int problemCount,
+        Difficulty difficulty
+) {
+}
