+++ src/main/java/com/ll/quizzle/domain/quiz/dto/submission/QuizSubmissionResultDTO.java
@@ -1,4 +1,4 @@
-package com.ll.quizzle.domain.quiz.dto;
+package com.ll.quizzle.domain.quiz.dto.submission;
 
 public record QuizSubmissionResultDTO(
         int questionNumber,
