+++ src/main/java/com/ll/quizzle/domain/quiz/service/RedisQuizAnswerService.java
@@ -1,49 +1,3 @@
-//package com.ll.quizzle.domain.quiz.service;
-//
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.data.redis.core.RedisTemplate;
-//import org.springframework.stereotype.Service;
-//
-//import java.time.Duration;
-//import java.util.LinkedHashMap;
-//import java.util.Map;
-//import java.util.UUID;
-//
-//@Service
-//public class RedisQuizAnswerService {
-//
-//    @Autowired
-//    private RedisTemplate<String, Object> redisTemplate;
-//
-//    private static final Duration QUIZ_TTL = Duration.ofMinutes(30);
-//
-//    public String saveQuiz(String quizText, Map<Integer, String> answerMap) {
-//        String quizId = UUID.randomUUID().toString();
-//        String redisKey = "quiz:" + quizId;
-//
-//        // 1. 퀴즈 텍스트는 hash에 저장 (순서는 의미없음)
-//        Map<String, String> redisMap = new LinkedHashMap<>();
-//        redisMap.put("quizText", quizText);
-//        redisTemplate.opsForHash().putAll(redisKey, redisMap);
-//        redisTemplate.expire(redisKey, QUIZ_TTL);
-//
-//        // 2. 문제 번호와 정답은 별도의 List에 순서대로 저장
-//        // 예: "1:b", "2:c", "3:a", ... 순서대로 저장됨
-//        String answerListKey = redisKey + ":answers";
-//        answerMap.entrySet().stream()
-//                .sorted(Map.Entry.comparingByKey()) // 숫자 오름차순 정렬
-//                .forEachOrdered(entry -> {
-//                    String entryValue = entry.getKey() + ":" + entry.getValue();
-//                    // 오른쪽에 추가하여 입력 순서를 유지
-//                    redisTemplate.opsForList().rightPush(answerListKey, entryValue);
-//                });
-//        redisTemplate.expire(answerListKey, QUIZ_TTL);
-//
-//        return quizId;
-//    }
-//}
-
-
 package com.ll.quizzle.domain.quiz.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
