+++ src/main/java/com/ll/quizzle/global/exceptions/handler/GlobalExceptionHandler.java
@@ -10,37 +10,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-//
-//@ControllerAdvice
-//@RequiredArgsConstructor
-//public class GlobalExceptionHandler {
-//
-//    @ExceptionHandler(ServiceException.class)
-//    public ResponseEntity<String> handle(ServiceException ex) {
-//
-//        String message = ex.getMsg();
-//
-//        return ResponseEntity
-//                .status(ex.getResultCode())
-//                .body(message);
-//    }
-//
-//    @ExceptionHandler(MethodArgumentNotValidException.class)
-//    public ResponseEntity<String> handleValidationException(MethodArgumentNotValidException ex) {
-//        List<FieldError> fieldErrors = ex.getBindingResult().getFieldErrors();
-//
-//        StringBuilder errorMessages = new StringBuilder();
-//        for (FieldError fieldError : fieldErrors) {
-//            errorMessages.append(fieldError.getField())
-//                    .append(": ")
-//                    .append(fieldError.getDefaultMessage())
-//                    .append("\n");
-//        }
-//
-//        return ResponseEntity.badRequest().body(errorMessages.toString());
-//    }
-//
-//}
+
 @ControllerAdvice
 @RequiredArgsConstructor
 public class GlobalExceptionHandler {
